requirements.txt
numpy==1.23.0
opencv-python==4.6.0.66
Pillow==9.2.0
qrcode==7.3.1
flask==2.1.2
python-dotenv==0.20.0
Flask-Cors==3.0.10